Move reboot syscall to system module
