+++ src/backend/linux_raw/system/syscalls.rs
@@ -7,10 +7,14 @@
 #![allow(clippy::undocumented_unsafe_blocks)]
 
 use super::types::RawUname;
-use crate::backend::conv::{ret, ret_infallible, slice};
+use crate::backend::{
+    c,
+    conv::{c_int, ret, ret_infallible, slice},
+};
 use crate::io;
 use crate::system::Sysinfo;
 use core::mem::MaybeUninit;
+use crate::system::RebootCommand;
 
 #[inline]
 pub(crate) fn uname() -> RawUname {
@@ -34,4 +38,16 @@ pub(crate) fn sysinfo() -> Sysinfo {
 pub(crate) fn sethostname(name: &[u8]) -> io::Result<()> {
     let (ptr, len) = slice(name);
     unsafe { ret(syscall_readonly!(__NR_sethostname, ptr, len)) }
+}
+
+#[inline]
+pub(crate) fn reboot(cmd: RebootCommand) -> io::Result<()> {
+    unsafe {
+        ret(syscall_readonly!(
+            __NR_reboot,
+            c_int(c::LINUX_REBOOT_MAGIC1),
+            c_int(c::LINUX_REBOOT_MAGIC2),
+            c_int(cmd as i32)
+        ))
+    }
 }
